feat : LadderString 생성, Name, Reward로 확장
