+++ src/main/java/ladder/domain/Reward.java
@@ -1,0 +1,9 @@
+package ladder.domain;
+
+public class Reward extends LadderString {
+
+    public Reward(String value) {
+        super(value);
+    }
+
+}
